fix(web-hosting): dashboard tile improvement
fix date issue
rename translations

resolves MBE-240
