+++ client/app/components/service-status/service-status-action.html
@@ -2,30 +2,25 @@
     <dl class="oui-tile__definition">
         <dt class="oui-tile__term">
             <div data-ng-if="!$ctrl.isExpired() && $ctrl.isInAutoRenew() && $ctrl.shouldDeleteAtExpiration()">
-                <span data-translate="expiration_resiliation_date"></span>
+                <span data-translate="expiration_resiliation_date_label"></span>
             </div>
             <div data-ng-if="!$ctrl.isExpired() && $ctrl.isInAutoRenew() && !$ctrl.shouldDeleteAtExpiration()">
-                <span data-translate="expiration_renew_date"></span>
+                <span data-translate="expiration_renew_date_label"></span>
             </div>
             <div data-ng-if="$ctrl.isExpired() || !$ctrl.isInAutoRenew()">
-                <span data-translate="{{ ::$ctrl.isExpired() ? 'expiration_resiliation_date' : 'expiration_expiration_date' }}"></span>
+                <span data-translate="{{ ::$ctrl.isExpired() ? 'expiration_resiliation_date_label' : 'expiration_expiration_date_label' }}"></span>
             </div>
         </dt>
         <dd class="oui-tile__description">
-            <div data-ng-if="!$ctrl.shouldDeleteAtExpiration()">
+            <div data-ng-if="!$ctrl.shouldDeleteAtExpiration() && $ctrl.isInAutoRenew()">
                 <span data-ng-bind=" ::$ctrl.getDate() | date:'MMM yyyy'"></span>
             </div>
-            <div data-ng-if="$ctrl.shouldDeleteAtExpiration() || $ctrl.isExpired()">
+            <div data-ng-if="$ctrl.shouldDeleteAtExpiration() || $ctrl.isExpired() || !$ctrl.isInAutoRenew()">
                 <span data-ng-bind=" ::$ctrl.getDate() | date:'mediumDate'"></span>
             </div>
         </dd>
     </dl>
     <oui-action-menu data-align="end" data-compact>
-        <oui-action-menu-item
-            data-aria-label="{{ ::'expiration_expiration_delete_cancel_label' | translate }}"
-            data-href="{{ $ctrl.getAutoRenewUrl() }}">
-            <span data-translate="expiration_expiration_delete_cancel_label"></span>
-        </oui-action-menu-item>
         <oui-action-menu-item
             data-ng-if="$ctrl.isExpired() || !$ctrl.isInAutoRenew()"
             data-aria-label="{{ ::'expiration_wizard_renew' | translate }}"
@@ -35,16 +30,19 @@
         </oui-action-menu-item>
         <oui-action-menu-item
             data-ng-if="!$ctrl.isExpired() && $ctrl.isInAutoRenew() && $ctrl.shouldDeleteAtExpiration()"
-            data-aria-label="{{ 'expiration_resiliation_cancel_label' | translate }}"
+            data-aria-label="{{ ::'expiration_resiliation_cancel_label' | translate }}"
             data-ng-href="{{ $ctrl.getAutoRenewUrl() }}">
             <span data-translate="expiration_resiliation_cancel_label"></span>
         </oui-action-menu-item>
-        <oui-action-menu-divider></oui-action-menu-divider>
         <oui-action-menu-item
-            data-aria-label="{{ ::'common_manage' | translate }}"
-            data-href="{{ $ctrl.getAutoRenewUrl() }}"
-            data-external>
+            data-aria-label="{{ ::'global_renew_management' | translate }}"
+            data-href="{{ $ctrl.getAutoRenewUrl() }}">
             <span data-translate="common_manage"></span>
+        </oui-action-menu-item>
+        <oui-action-menu-item
+            data-aria-label="{{ ::'expiration_expiration_delete_cancel_label' | translate }}"
+            data-href="{{ $ctrl.getAutoRenewUrl() }}">
+            <span data-translate="expiration_expiration_delete_cancel_label"></span>
         </oui-action-menu-item>
     </oui-action-menu>
 </div>
