+++ client/app/hosting/general-informations/GENERAL_INFORMATIONS.controller.js
@@ -182,7 +182,7 @@ angular.module('App').controller(
       this.$scope.setAction('change-main-domain/hosting-change-main-domain', this.$scope.hosting);
     }
 
-    isOffer() {
+    isHostingOffer() {
       return _.includes(['KIMSUFI_2015', '__60_FREE', 'DEMO_1_G', 'START_1_M', 'START_10_M', '_ASPFREE'], this.$scope.hosting.offer);
     }
   },
