+++ client/app/hosting/general-informations/GENERAL_INFORMATIONS.html
@@ -2,7 +2,7 @@
 
     <oui-spinner data-ng-if="!hosting"></oui-spinner>
 
-    <div data-ng-show="hosting">
+    <div data-ng-if="hosting">
         <div data-ovh-alert="{{alerts.main}}"></div>
 
         <div class="alert alert-dismissible alert-info" role="alert"
@@ -124,7 +124,7 @@
                             <span class="word-break" data-ng-bind="::hosting.serviceName"></span>
                         </oui-tile-description>
                         <oui-action-menu data-compact data-align="end"
-                                         data-ng-if="$ctrl.isOffer()">
+                                         data-ng-if="$ctrl.isHostingOffer()">
                             <oui-action-menu-item data-on-click="$ctrl.changeMainDomain()">
                                 <span data-translate="hosting_dashboard_service_main_domain_change"></span>
                             </oui-action-menu-item>
