+++ client/app/hosting/translations/Messages_fr_FR.xml
@@ -1528,14 +1528,11 @@
    <translation id="common_subscription_contacts_admin">Administrateur</translation>
    <translation id="common_subscription_contacts_billing">Facturation</translation>
    <translation id="common_subscription_contacts_tech">Technique</translation>
-   <translation id="core_service_expiration_auto_renew">Renouvellement automatique prévu en</translation>
 
-   <translation id="expiration_new_tab">S'ouvre dans un nouvel onglet</translation>
-   <translation id="expiration_renew_date">Renouvellement automatique prévu en </translation>
+   <translation id="expiration_renew_date_label">Renouvellement automatique prévu en </translation>
    <translation id="expiration_resiliation_cancel_label">Annuler la résiliation</translation>
    <translation id="expiration_wizard_renew">Renouveler</translation>
-   <translation id="expiration_resiliation_date">Résiliation prévue le</translation>
-   <translation id="expiration_expiration_date">Expire le</translation>
+   <translation id="expiration_resiliation_date_label">Résiliation prévue le</translation>
+   <translation id="expiration_expiration_date_label">Expire le</translation>
    <translation id="expiration_expiration_delete_cancel_label">Supprimer / Résilier mon hébergement</translation>
-
 </translations>
